rename #41 : 폴더 관심사 분리 및 불필요한 파일 삭제
